backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
PyJWT==2.8.0
python-dotenv==1.0.0
pandas==2.1.4
geopy==2.4.1